+++ release/plugins/jsx/hide/index.es6.js
@@ -1,0 +1,1 @@
+export { hide } from './hide.es6.js';