+++ release/plugins/jsx/hide/index.js
@@ -1,0 +1,9 @@
+'use strict';
+
+Object.defineProperty(exports, '__esModule', { value: true });
+
+var hide = require('./hide.js');
+
+
+
+exports.hide = hide.hide;