+++ release/plugins/jsx/index.d.ts
@@ -4,3 +4,5 @@ export * from './router';
 export * from './link';
 export * from './delay';
 export * from './show';
+export * from './hide';
+export * from './switch';