+++ release/plugins/jsx/index.es6.js
@@ -4,3 +4,5 @@ import './router/index.es6.js';
 import './link/index.es6.js';
 import './delay/index.es6.js';
 import './show/index.es6.js';
+import './hide/index.es6.js';
+import './switch/index.es6.js';