+++ release/plugins/jsx/index.js
@@ -6,4 +6,6 @@ require('./router/index.js');
 require('./link/index.js');
 require('./delay/index.js');
 require('./show/index.js');
+require('./hide/index.js');
+require('./switch/index.js');