+++ release/plugins/jsx/show/show.d.ts
@@ -1,6 +1,6 @@
 import { type JSXPluginElement } from '@innet/jsx';
 import { type StateProp } from '../../../types';
 export interface ShowProps {
-    state: StateProp<any>;
+    when: StateProp<any>;
 }
-export declare function show({ props: { state }, children }: JSXPluginElement<ShowProps>, handler: any): any;
+export declare function show({ props: { when }, children }: JSXPluginElement<ShowProps>, handler: any): any;