+++ release/plugins/jsx/switch/index.es6.js
@@ -1,0 +1,1 @@
+export { switchPlugin } from './switch.es6.js';