+++ release/plugins/jsx/switch/index.js
@@ -1,0 +1,9 @@
+'use strict';
+
+Object.defineProperty(exports, '__esModule', { value: true });
+
+var _switch = require('./switch.js');
+
+
+
+exports.switchPlugin = _switch.switchPlugin;