+++ release/plugins/jsx/switch/switch.js
@@ -2,6 +2,7 @@
 
 Object.defineProperty(exports, '__esModule', { value: true });
 
+var jsx = require('@innet/jsx');
 var innet = require('innet');
 require('../../../utils/index.js');
 var inject = require('../../../utils/inject/inject.js');
@@ -10,8 +11,9 @@ function _interopDefaultLegacy (e) { return e && typeof e === 'object' && 'defau
 
 var innet__default = /*#__PURE__*/_interopDefaultLegacy(innet);
 
-function show({ props: { when }, children }, handler) {
-    return innet__default["default"](inject.inject(when, state => state ? children : null), handler);
+function switchPlugin({ props: { of }, children }, handler) {
+    const slots = jsx.getSlots(handler, children);
+    return innet__default["default"](inject.inject(of, state => { var _a; return (_a = slots[state]) !== null && _a !== void 0 ? _a : slots['']; }), handler);
 }
 
-exports.show = show;
+exports.switchPlugin = switchPlugin;